change submitter_id from integer to bigint

+++ db/migrate/20181105145906_add_submitter_id_to_users.rb
@@ -1,6 +1,6 @@
 class AddSubmitterIdToUsers < ActiveRecord::Migration[5.2]
   def change
-    add_column :users, :submitter_id, :integer
+    add_column :users, :submitter_id, :bigint
     add_index :users, :submitter_id, unique: true
   end
 end
